refactor code to make it more modular
